att corrigindo dados int

+++ src/homework_05/Homework_05.java
@@ -35,9 +35,9 @@ public class Homework_05 {
     /*-----------------------------*/
     private static void executarCalculo() {
         System.out.println("\nDigite o primeiro valor para calcular:");
-        byte valor1 = leia.nextByte();
+        int valor1 = leia.nextByte();
         System.out.println("\nDigite o segundo valor para calcular:");
-        byte valor2 = leia.nextByte();
+        int valor2 = leia.nextByte();
         System.out.println("\n" + calculos(valor1, valor2));
         novoCalculo();
     }
